9-research-assistant-agent/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "research-assistant-agent"
version = "0.1.0"
description = "AI Research Assistant for comprehensive research using search, analysis, and synthesis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "research_assistant"}]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "1.16.0"
google-generativeai = "^0.8.5"
python-dotenv = "^1.1.0"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.42.0"}
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
python-multipart = "^0.0.6"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.1.0"

[tool.poetry.scripts]
deploy-local = "deployment.local:main"
deploy-remote = "deployment.remote:main"
